+++ src/components/gtt-client/index.ts
@@ -1,6 +1,10 @@
 // This line exports all the members from the redmine module file.
 export * from './redmine';
 
+// import 'ol/ol.css';
+// import 'ol-ext/dist/ol-ext.min.css';
+// import 'ol-ext/filter/Base';
+
 // This line imports the GttClient class from the gtt-client-class module file and then re-exports it as the default export of this module file.
-import GttClient from './gtt-client-class';
+import GttClient from './GttClient';
 export { GttClient };
