+++ src/components/gtt-client/init/contents.ts
@@ -1,0 +1,3 @@
+export function initContents(target: HTMLElement): DOMStringMap {
+  return target.dataset;
+}
